==> 00_Installation.Rmd <==
---
title: "Practical: Installation"
subtitle: "Transcriptome Analysis Workshop"
author: "Adam Cribbs"
date: "24/01/2022"
output:
  html_document:
    theme: cosmo
    toc: yes
---

```{r, out.width = "40%", echo=FALSE}
htmltools::img(src = knitr::image_uri(file.path("logo.png")), 
               alt = 'logo', 
               style = 'position:absolute; top:0; right:0; padding:10px;',
               width='300')

```


Before you run the tutorial, you will need to install the following packages. I have written a few helper functions that will facilitate the installation process:

```{r, message=FALSE}
# Install CRAN packages
check.packages <- function(pkg){
    new.pkg <- pkg[!(pkg %in% installed.packages()[, "Package"])]
    if (length(new.pkg)) 
        install.packages(new.pkg, dependencies = TRUE)
    sapply(pkg, require, character.only = TRUE)
}

packages<-c("Seurat", "tidyverse", "BiocManager", "Matrix", "cowplot", "ggplot2",
            "scales", "devtools", "reticulate")
check.packages(packages)


# Install bioconductor packages
check_biocmanager <- function(pkg){
  if (!requireNamespace(pkg, quietly = TRUE))
    BiocManager::install(pkg)
  sapply(pkg, require, character.only = TRUE)}

bioc_packages<-c("tximport", "scater", "limma", "EnsDb.Hsapiens.v75",
            "AnnotationHub", "ensembldb", "SingleCellExperiment", "scClassify", "SC3", "monocle3",
            'DelayedArray', 'DelayedMatrixStats', 'org.Hs.eg.db', 'org.Mm.eg.db', 'fishpond')
check_biocmanager(bioc_packages)

# these packages are not currently on any package manager so you will need to
# install these individually from github using devtools (installed above).

# devtools::install_github("cole-trapnell-lab/garnett")


```
